app/requirements.txt
Django==5.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
django-modeltranslation==0.19.12
django-environ==0.11.2
django-imagekit==5.0.0
django-mptt==0.16.0
djoser==2.3.1
drf-spectacular==0.28.0
django-unfold==0.71.0
django-phonenumber-field[phonenumbers]==8.0.0
phonenumbers==9.0.7
qrcode==8.0
reportlab==4.2.5
PyPDF2==3.0.0
django-cachalot==2.7.0
django-redis==5.4.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
uvicorn==0.34.0
websockets==14.2
Pillow==10.4.0
psycopg2-binary~=2.9.10
gunicorn==23.0.0
unidecode==1.3.8
python-telegram-bot==21.11.0
django-silk==5.3.2
paho-mqtt==2.1.0
deep-translator==1.11.4
django-axes==8.0.0
django-import-export==4.3.7
openpyxl==3.1.5
drf-orjson-renderer==1.7.3
openai==2.3.0
sentry-sdk[django]==2.26.1
